scheduler cleanup
