+++ composer/optim/__init__.py
@@ -10,15 +10,15 @@ from composer.optim.optimizer_hparams import OptimizerHparams as OptimizerHparam
 from composer.optim.optimizer_hparams import RAdamHparams as RAdamHparams
 from composer.optim.optimizer_hparams import RMSPropHparams as RMSPropHparams
 from composer.optim.optimizer_hparams import SGDHparams as SGDHparams
-from composer.optim.scheduler import ConstantLRHparams as ConstantLRHparams
-from composer.optim.scheduler import CosineAnnealingLRHparams as CosineAnnealingLRHparams
-from composer.optim.scheduler import CosineAnnealingWarmRestartsHparams as CosineAnnealingWarmRestartsHparams
-from composer.optim.scheduler import CosineAnnealingWithWarmupLRHparams as CosineAnnealingWithWarmupLRHparams
-from composer.optim.scheduler import ExponentialLRHparams as ExponentialLRHparams
-from composer.optim.scheduler import LinearLRHparams as LinearLRHparams
-from composer.optim.scheduler import LinearWithWarmupLRHparams as LinearWithWarmupLRHparams
-from composer.optim.scheduler import MultiStepLRHparams as MultiStepLRHparams
-from composer.optim.scheduler import MultiStepWithWarmupLRHparams as MultiStepWithWarmupLRHparams
-from composer.optim.scheduler import PolynomialLRHparams as PolynomialLRHparams
-from composer.optim.scheduler import SchedulerHparams as SchedulerHparams
-from composer.optim.scheduler import StepLRHparams as StepLRHparams
+from composer.optim.scheduler_hparams import ConstantLRHparams as ConstantLRHparams
+from composer.optim.scheduler_hparams import CosineAnnealingLRHparams as CosineAnnealingLRHparams
+from composer.optim.scheduler_hparams import CosineAnnealingWarmRestartsHparams as CosineAnnealingWarmRestartsHparams
+from composer.optim.scheduler_hparams import CosineAnnealingWithWarmupLRHparams as CosineAnnealingWithWarmupLRHparams
+from composer.optim.scheduler_hparams import ExponentialLRHparams as ExponentialLRHparams
+from composer.optim.scheduler_hparams import LinearLRHparams as LinearLRHparams
+from composer.optim.scheduler_hparams import LinearWithWarmupLRHparams as LinearWithWarmupLRHparams
+from composer.optim.scheduler_hparams import MultiStepLRHparams as MultiStepLRHparams
+from composer.optim.scheduler_hparams import MultiStepWithWarmupLRHparams as MultiStepWithWarmupLRHparams
+from composer.optim.scheduler_hparams import PolynomialLRHparams as PolynomialLRHparams
+from composer.optim.scheduler_hparams import SchedulerHparams as SchedulerHparams
+from composer.optim.scheduler_hparams import StepLRHparams as StepLRHparams
